% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XETAfunctions.R
\name{ana}
\alias{ana}
\title{ana: Conducting analysis of XETA}
\usage{
ana(dataframe, mixedaov, dagostino = FALSE, seed = 123)
}
\arguments{
\item{dataframe}{A dataframe containing only the original spiked or unspiked
samples without T4 control.}

\item{mixedaov}{A mixed ANOVA object with Fluor as dependent variable and
 Conc + (1|Replicate)+(1|Replicate:Conc)+(1|Replicate:Conc:Row) as linear
predictor. The last random term is optional.}

\item{dagostino}{boolean. Default is FALSE; when FALSE, the shapiro wilks
test from the stats package is carried out, when TRUE the dAgostina Pearson
normality test from the fBasics test is carried out.}

\item{seed}{numeric. The number use to set the seed. Default is 123}
}
\value{
A list (collection) containing the trimmed or untrimmed data frame,
  two summary data frames containing information summarized based on the
  concentration and the replicate, the result from the shapiro and levene
  test, the boxplots per replicate, the result of the monotonicity test as
well as the result from the Dunnett's and Williams test.
}
\description{
This function is used by the data_prep function
 (which decides whether trimming and/or transformations are conducted) to
 carry out the actual analysis.
However, the ana function can also be used independently from data_prep,
in case a different approach than recommended by data_prep is anticipated.
}
\examples{
mixedaov <- lme4::lmer(
  Fluor ~ Conc +
    (1 | Replicate) + (1 | Replicate:Conc) + (1 | Replicate:Conc:Row),
  data = xeredar::testDataSpiked, REML = TRUE
)
Result1 <- ana(xeredar::testDataSpiked, mixedaov)

# second example:
mixedaov <- lme4::lmer(Fluor ~ Conc + (1 | Replicate) + (1 | Replicate:Conc),
  data = xeredar::valid_data_xeta[["triac_france_spiked"]], REML = TRUE
)
Result2 <- ana(xeredar::valid_data_xeta[["triac_france_spiked"]], mixedaov)

# third example:
mixedaov <- lme4::lmer(log(Fluor) ~ Conc + (1 | Replicate) +
  (1 | Replicate:Conc), data =
  xeredar::valid_data_xeta[["triac_japan_spiked"]],
  REML = TRUE
)
Result2 <- ana(xeredar::valid_data_xeta[["triac_japan_spiked"]], mixedaov)

}
