% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XETAfunctions.R
\name{anova_assumpts}
\alias{anova_assumpts}
\title{anova_assumpts checks whether normality and variance homogeneity
assumptions are fulfilled.}
\usage{
anova_assumpts(dagostino, mixedaov, alpha)
}
\arguments{
\item{dagostino}{boolean. Default is FALSE; when FALSE, the shapiro wilks
test from the stats package is carried out, when TRUE the dAgostino Pearson
normality test from the fBasics test is carried out.}

\item{mixedaov}{a mixed anova model object specified with lmer
from lme4.}

\item{alpha}{Numeric. Default is 0.01. This value is the alpha value for
the normality and variance homogeneity test.}
}
\value{
boolean; if TRUE, the p-values from the normality and
 variance homogeneity test are larger then alpha. If FALSE, one
 or both p-values from those tests is/are smaller than alpha.
}
\description{
anova_assumpts checks whether normality and variance homogeneity
assumptions are fulfilled.
}
\examples{
mixedaov <- lme4::lmer(Fluor ~ Conc + (1 | Replicate) + (1 | Replicate:Conc),
  data = xeredar::valid_data_xeta[["triac_france_unspiked"]], REML = TRUE
)
anova_assumpts(dagostino = TRUE, mixedaov = mixedaov, alpha = 0.01)
}
