% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XETAfunctions.R
\name{data_prep}
\alias{data_prep}
\title{data_prep: This function recommends a data preparation method
(transformation and/or trimming) and conducts the analysis
accordingly.}
\usage{
data_prep(
  dataframe,
  row = TRUE,
  dagostino = FALSE,
  trimming = TRUE,
  outlier = TRUE,
  boxcox = TRUE,
  artc = FALSE,
  alpha = 0.01
)
}
\arguments{
\item{dataframe}{A dataframe containing only the original spiked or unspiked
samples without T4 control.}

\item{row}{Please specify FALSE (boolean) in case you did not record the
exposure well ID/ Row of the 96 well plate or you want to use the reduced
mixed ANOVA model where the exposure well ID is not a random effect.}

\item{dagostino}{boolean. Default is FALSE; when FALSE, the shapiro
wilks test from the stats package is carried out, when TRUE the
dAgostino Pearson normality test from the fBasics test is carried out.}

\item{trimming}{boolean. When TRUE, trimming is conducted when required. When
FALSE trimming is never conducted. Default is TRUE.}

\item{outlier}{boolean, When TRUE, outlier removal is conducted when
required. When FALSE, outlier removal is never conducted.
Default is TRUE.}

\item{boxcox}{boolean, When TRUE boxcox transformation is performed when
necessary and when no zeros are present in the data. When FALSE it will never
be conducted. Default is TRUE.}

\item{artc}{boolean, When TRUE ART-C transformation is performed when
necessary.
When FALSE it will never be conducted. Default is FALSE.
This feature is experimental!}

\item{alpha}{Numeric. Default is 0.01. This value is the alpha value for the
normality and variance homogeneity test.}
}
\value{
A list (collection) containing the trimmed or untrimmed data frame,
two summary data frames containing information summarized based on the
concentration and the replicate, the result from the shapiro and levene test,
the boxplots per replicate, the result of the monotonicity test as well
as the result from the Dunnett's and Williams test. Different from the
output of the ana function, the returned list also contains the mixed
Anova model (mixedaov) whereby Fluor is transformed as recommended
and the trimmed or untrimmed data was utilized. Furthermore, the output
also contains a reasoning for the recommended transformation
and trimming.
}
\description{
This function decides whether trimming, outlier removal and/or
transformations are conducted. The actual analysis is carried out by the
ana function.
}
\examples{

# XETA example with thyroid active chemical (data from validation studies):
result1_1 <- data_prep(xeredar::valid_data_xeta[["triac_japan_unspiked"]],
 row = FALSE)
result1_2 <- data_prep(xeredar::valid_data_xeta[["triac_japan_spiked"]],
 row = FALSE)
# third example:
result2_1 <- data_prep(xeredar::valid_data_xeta[["triac_france_spiked"]],
 row = FALSE)
result2_2 <- data_prep(xeredar::valid_data_xeta[["triac_france_unspiked"]],
 row = FALSE)
# fourth example:
result3_1 <- data_prep(xeredar::valid_data_xeta[["triac_usa_spiked"]],
 row = FALSE)
result3_2 <- data_prep(xeredar::valid_data_xeta[["triac_usa_unspiked"]],
 row = FALSE)

# RADAR example androgen axis active chemical (data from validation studies)
result4_1 <- data_prep(
xeredar::RADAR_valid_data_table_spiked_unspiked[["AnastrozoleCEFAS_Unspiked"]],
  row = FALSE, dagostino = TRUE, trimming = TRUE, outlier = TRUE,
  boxcox = FALSE, artc = FALSE, alpha = 0
)
result4_2 <- data_prep(
xeredar::RADAR_valid_data_table_spiked_unspiked[["AnastrozoleCEFAS_Spiked"]],
  row = FALSE, dagostino = TRUE, trimming = FALSE, outlier = FALSE,
  boxcox = FALSE, alpha = 0
)
result5_1 <- data_prep(
xeredar::RADAR_valid_data_table_spiked_unspiked[["AnastrozoleFIWI_Unspiked"]],
  row = FALSE,
  dagostino = TRUE, trimming = TRUE, outlier = FALSE,
  boxcox = FALSE, artc = FALSE, alpha = 0.05
)
result5_2 <- data_prep(
xeredar::RADAR_valid_data_table_spiked_unspiked[["AnastrozoleFIWI_Spiked"]],
  row = FALSE,
  dagostino = TRUE, trimming = FALSE
)
result6_1 <- data_prep(
xeredar::RADAR_valid_data_table_spiked_unspiked[["AnastrozoleFIWI_Unspiked"]],
  row = FALSE,
  dagostino = TRUE, trimming = FALSE
)
result6_2 <- data_prep(
xeredar::RADAR_valid_data_table_spiked_unspiked[["AnastrozoleIDEA_Spiked"]],
  row = FALSE,
  dagostino = TRUE, trimming = FALSE
)
result7_1 <- data_prep(
xeredar::RADAR_valid_data_table_spiked_unspiked[["AnastrazoleWATCHFR_Unspiked"]],
  row = FALSE, dagostino = TRUE, trimming = FALSE
)
result7_2 <- data_prep(
xeredar::RADAR_valid_data_table_spiked_unspiked[["AnastrazoleWATCHFR_Spiked"]],
  row = FALSE,
  dagostino = TRUE, trimming = FALSE
)
}
