% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XETAfunctions.R
\name{data_summary}
\alias{data_summary}
\title{Function for summary of results}
\usage{
data_summary(datalist)
}
\arguments{
\item{datalist}{the list that is provided by the
analysis of either spiked or unspiked data}
}
\value{
A summary file
}
\description{
Function for summary of results
}
\examples{
Result1_1 <- validity(xeredar::valid_data_xeta[["triac_japan_combined"]],
  xeredar::valid_data_xeta[["triac_japan_unspiked"]],
  row = FALSE
)
data_summary(Result1_1)

Result1_2 <- validity(xeredar::valid_data_xeta[["triac_japan_combined"]],
  xeredar::valid_data_xeta[["triac_japan_spiked"]],
  row = FALSE, alpha = 0.005
)
data_summary(Result1_2)

Result2_1 <- validity(xeredar::valid_data_xeta[["triac_france_combined"]],
  xeredar::valid_data_xeta[["triac_france_unspiked"]],
  row = FALSE
)
data_summary(Result2_1)

Result2_2 <- data_prep(dataframe =
xeredar::valid_data_xeta[["triac_france_spiked"]], row = FALSE)
data_summary(Result2_2)

Result3_1 <- data_prep(dataframe =
xeredar::valid_data_xeta[["t4_japan_spiked"]], row = FALSE)
data_summary(Result3_1)

Result3_2 <- data_prep(dataframe =
xeredar::valid_data_xeta[["ptu_usa_spiked"]], row = FALSE)
data_summary(Result3_2)
}
