% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eventTimeData}
\alias{eventTimeData}
\title{eventTimeData from StatCharrms}
\format{
An object of class \code{data.frame} with 240 rows and 4 columns.
}
\usage{
eventTimeData
}
\description{
The same dataset as stored in the StatCharrms package
}
\keyword{datasets}
