% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fecundityData}
\alias{fecundityData}
\title{fecundityData from StatCharrms}
\format{
An object of class \code{data.frame} with 1453 rows and 5 columns.
}
\usage{
fecundityData
}
\description{
The same dataset as stored in the StatCharrms package
}
\keyword{datasets}
