% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCharrms_original.R
\name{getLineContrast}
\alias{getLineContrast}
\title{getLineContrast}
\usage{
getLineContrast(Data, Treatment)
}
\arguments{
\item{Data}{A data set.}

\item{Treatment}{The name (as a string) of the treatment variable.}
}
\value{
Returns a numeric vector of contrasts.
}
\description{
This is a function originally encoded by Joe Swintek
 for the R package StatCharrms. Generates linear contrasts for
 testing monotonicity based on the OECD guidance No. 54 for statistics
 in ecotoxicology.
}
\references{
\url{https://www.rdocumentation.org/packages/StatCharrms/versions/0.90.96}

OECD (2006) Current approaches in the statistical analysis of
ecotoxicity data: A guidance to application. Environmental Health and
Safety Publications. Series on Testing and Assessment. No. 54. Paris.
}
\author{
Joe Swintek
}
