% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lengthWeightData}
\alias{lengthWeightData}
\title{lengthWeightData from StatCharrms}
\format{
An object of class \code{data.frame} with 370 rows and 9 columns.
}
\usage{
lengthWeightData
}
\description{
The same dataset as stored in the StatCharrms package
}
\keyword{datasets}
