% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MeansAllStudiesperConcUnspiked}
\alias{MeansAllStudiesperConcUnspiked}
\title{Mean values for each concentration group, compound and concentration for the power analysis 
based on REACTIV validation data. Based on these mean values, the simulation is carried out.}
\format{
An object of class \code{data.frame} with 2772 rows and 10 columns.
}
\usage{
MeansAllStudiesperConcUnspiked
}
\description{
Mean values for each concentration group, compound and concentration for the power analysis 
based on REACTIV validation data. Based on these mean values, the simulation is carried out.
}
\keyword{datasets}
