% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Negative_Amant_Watchfrog_RADAR_Unspiked}
\alias{Negative_Amant_Watchfrog_RADAR_Unspiked}
\title{Negative_Amant_Watchfrog_RADAR_Unspiked data used for testing the functions.
It contains only unspiked data.The column Conc is a factor with specified
levels, as the Williams test will otherwise not order the concentrations in
the correct way. This dataset is from the RADAR study validation in the lab
Watchfrog with an inert chemical .}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 378 rows and 4 columns.
}
\usage{
Negative_Amant_Watchfrog_RADAR_Unspiked
}
\description{
Negative_Amant_Watchfrog_RADAR_Unspiked data used for testing the functions.
It contains only unspiked data.The column Conc is a factor with specified
levels, as the Williams test will otherwise not order the concentrations in
the correct way. This dataset is from the RADAR study validation in the lab
Watchfrog with an inert chemical .
}
\keyword{datasets}
