% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XETAfunctions.R
\name{norm_test}
\alias{norm_test}
\title{norm_test choose whether you want to carry out the d'Agostino Pearson
normality test from the fBasics package or the shapiro wilks test from the
stats package}
\usage{
norm_test(dagostino = FALSE, mixedaov)
}
\arguments{
\item{dagostino}{boolean. Default is FALSE; when FALSE, the shapiro wilks
test from the stats package is carried out, when TRUE the dAgostino Pearson
normality test from the fBasics test is carried out.}

\item{mixedaov}{a mixed anova model object specified with lmer
from lme4.}
}
\value{
a list with p-values from the selected normality test.
}
\description{
This function is called from the Ana and dataPrep function
}
\examples{
mixedaov <- lme4::lmer(Fluor ~ Conc + (1 | Replicate) + (1 | Replicate:Conc),
  data = xeredar::valid_data_xeta[["triac_france_spiked"]], REML = TRUE
)
norm_test(dagostino = TRUE, mixedaov = mixedaov)
}
