% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Pos_Anas_IDEA_RADAR_Spiked}
\alias{Pos_Anas_IDEA_RADAR_Spiked}
\title{Pos_Anas_IDEA_RADAR_Spiked data used for testing the functions.
It contains only spiked data.The column Conc is a factor with specified
levels, as the Williams test will otherwise not order the concentrations in
the correct way. This dataset is from the RADAR study validation in the lab
IDEA with an androgen axis active chemical.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 360 rows and 4 columns.
}
\usage{
Pos_Anas_IDEA_RADAR_Spiked
}
\description{
Pos_Anas_IDEA_RADAR_Spiked data used for testing the functions.
It contains only spiked data.The column Conc is a factor with specified
levels, as the Williams test will otherwise not order the concentrations in
the correct way. This dataset is from the RADAR study validation in the lab
IDEA with an androgen axis active chemical.
}
\keyword{datasets}
