% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Pos_mDHT_Fraunhofer_RADAR}
\alias{Pos_mDHT_Fraunhofer_RADAR}
\title{Pos_mDHT_Fraunhofer_RADAR data used for testing the functions.
It contains spiked and unspiked data.The column Conc is a factor with
specified levels, as the Williams test will otherwise not order the
concentrations in the correct way. This dataset is from the RADAR study
validation in the lab Fraunhofer with an androgen axis active chemical.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 732 rows and 4 columns.
}
\usage{
Pos_mDHT_Fraunhofer_RADAR
}
\description{
Pos_mDHT_Fraunhofer_RADAR data used for testing the functions.
It contains spiked and unspiked data.The column Conc is a factor with
specified levels, as the Williams test will otherwise not order the
concentrations in the correct way. This dataset is from the RADAR study
validation in the lab Fraunhofer with an androgen axis active chemical.
}
\keyword{datasets}
