% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCharrms_original.R
\name{rankTransform}
\alias{rankTransform}
\title{rankTransform}
\usage{
rankTransform(Data, VecName)
}
\arguments{
\item{Data}{A dataset}

\item{VecName}{The name (as a string) of the response to be transformed.}
}
\value{
The original data set with a new variable called
'TransformedResponse' which is the rank transform of
 the response.
}
\description{
This is a function originally encoded by Joe Swintek for the
 R package StatCharrms. It Transforms a response by rank into critical values
 of the standard normal distribution. In the case of ties this function will
 use the mean of the transformed response.
 This function will rank transform the data the
 same way the SAS code below does:
 proc rank data=sorted out=assumtst normal=blom
 ties=mean;
 var Counts;
 by generation;
 ranks rank_EGGS;
 run;
}
\examples{
data(lengthWeightData)
SubData <- lengthWeightData[lengthWeightData$Age == "16 week", ]
SubData <- SubData[SubData$Generation == "F1", ]
SubData <- SubData[SubData$SEX == "M", ]
RankData <- rankTransform(Data = SubData, VecName = "WEIGHT")
head(RankData)
}
\references{
\url{https://www.rdocumentation.org/packages/StatCharrms/versions/0.90.96}
}
\author{
Joe Swintek
}
