% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reactiv_data}
\alias{reactiv_data}
\title{REACTIV data used for testing the functions.
It contains both spiked and unspiked data.The column Conc is a factor with specified
levels, as the Williams test will otherwise not order the concentrations in
the correct way. This dataset is an artificiald dataset.}
\format{
An object of class \code{data.frame} with 312 rows and 5 columns.
}
\usage{
reactiv_data
}
\description{
REACTIV data used for testing the functions.
It contains both spiked and unspiked data.The column Conc is a factor with specified
levels, as the Williams test will otherwise not order the concentrations in
the correct way. This dataset is an artificiald dataset.
}
\keyword{datasets}
