% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REACTIV_valid_data_table_spiked_unspiked}
\alias{REACTIV_valid_data_table_spiked_unspiked}
\title{REACTIV_valid_data_table_spiked_unspiked list of data used for testing the
functions.
It contains spiked and unspiked data from all substances and labs from the
REACTIV validation studies.
For all datasets the column Conc is a factor with specified
levels, as the Williams test will otherwise not order the concentrations in
the correct way.}
\format{
An object of class \code{list} of length 71.
}
\usage{
REACTIV_valid_data_table_spiked_unspiked
}
\description{
REACTIV_valid_data_table_spiked_unspiked list of data used for testing the
functions.
It contains spiked and unspiked data from all substances and labs from the
REACTIV validation studies.
For all datasets the column Conc is a factor with specified
levels, as the Williams test will otherwise not order the concentrations in
the correct way.
}
\keyword{datasets}
