% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resultsUnspiked_power_reactive}
\alias{resultsUnspiked_power_reactive}
\title{Results  of power analysis of simulated unspiked REACTIV data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 220 rows and 5 columns.
}
\usage{
resultsUnspiked_power_reactive
}
\description{
Results  of power analysis of simulated unspiked REACTIV data
}
\keyword{datasets}
