% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessData.R
\name{rm_outlier}
\alias{rm_outlier}
\title{rm_outlier Function to remove outliers based on the Tukey outlier rule.}
\usage{
rm_outlier(dataframe, row = TRUE)
}
\arguments{
\item{dataframe}{The untrimmed data frame with columns
"Treatment", "Replicate", "Fluor", "Row", "Well", "Conc", "Dose".
The values that were spiked must contain a "T3" in the Treatment name.
The T4 control must be named FETAXT3T4. Take care that you do not contain
any spaces in the treatment names!}

\item{row}{Please specify FALSE (boolean) in case you did not record the
exposure well ID/ Row of the 96 well plate or you want to use the reduced
mixed ANOVA model where the exposure well ID is not a random effect.}
}
\value{
Returns a list containing two data frames with the measurements
  identified as outliers (outliers) and another data frame containing the
  remaining measurements (wt_outlier).
}
\description{
rm_outlier Function to remove outliers based on the Tukey outlier rule.
}
