% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Simulation_Info}
\alias{Simulation_Info}
\title{Data containing information from variance analysis of original REACTIV validation studies. 
Based on this data, and on MeansAllStudiesperConcSpiked and MeansAllStudiesperConcUnspiked the simualted data is created.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 135 rows and 10 columns.
}
\usage{
Simulation_Info
}
\description{
Data containing information from variance analysis of original REACTIV validation studies. 
Based on this data, and on MeansAllStudiesperConcSpiked and MeansAllStudiesperConcUnspiked the simualted data is created.
}
\keyword{datasets}
