% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testDataUnspiked}
\alias{testDataUnspiked}
\title{testDataUnspiked data used for testing the functions.
This is a data frame containing the spiked measurements of a xeta studies.
The two test_data sets are the only datasets with exposure well information
(column Row). The column Conc in each dataframe is a factor with specified
levels, as the Williams test will otherwise not order the concentrations
in the correct way.}
\format{
An object of class \code{data.frame} with 240 rows and 5 columns.
}
\usage{
testDataUnspiked
}
\description{
testDataUnspiked data used for testing the functions.
This is a data frame containing the spiked measurements of a xeta studies.
The two test_data sets are the only datasets with exposure well information
(column Row). The column Conc in each dataframe is a factor with specified
levels, as the Williams test will otherwise not order the concentrations
in the correct way.
}
\keyword{datasets}
