% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessData.R
\name{trim}
\alias{trim}
\title{trim function}
\usage{
trim(dataframe, lwr = 0.1, upr = 0.9)
}
\arguments{
\item{dataframe}{A data frame with columns "Treatment", "Replicate", "Fluor"}

\item{lwr}{lower percentile treshold}

\item{upr}{upper percentile treshold}
}
\value{
Returns a list containing two data frames with the trimmed out
  measurements (trimmed_out) and the trimmed data (dat_trim)
}
\description{
Computes quantiles for each Treatment x Replicate combination
  and trim all observations with Flour greater than the quantile tresholds.
}
\examples{
trimmedDataList <- trim(xeredar::testDataSpiked)
}
