% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{valid_data_xeta}
\alias{valid_data_xeta}
\title{valid_data_xeta data used for testing the functions.
It is a list of data frames containing the spiked/unspiked or
combined measurements of all xeta validation studies.
The column conc in each dataframe is a factor with specified levels,
as the Williams test will otherwise not order the concentrations in the
correct way.}
\format{
An object of class \code{list} of length 108.
}
\usage{
valid_data_xeta
}
\description{
valid_data_xeta data used for testing the functions.
It is a list of data frames containing the spiked/unspiked or
combined measurements of all xeta validation studies.
The column conc in each dataframe is a factor with specified levels,
as the Williams test will otherwise not order the concentrations in the
correct way.
}
\keyword{datasets}
