% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validators.R
\name{validate_dfxeta}
\alias{validate_dfxeta}
\title{Valdidating XETA dataframe}
\usage{
validate_dfxeta(df, row = TRUE)
}
\arguments{
\item{df}{The dataframe that is used in the parent function}

\item{row}{logical. Please specify FALSE (boolean) in case you did not record
the exposure well ID/ Row of the 96 well plate or you want to use the
reduced mixed ANOVA model where the exposure well ID is not a random
effect.}
}
\value{
In case the dataframe is valid,
  the function invisibly returns it to the parent function
}
\description{
Valdidating XETA dataframe
}
\examples{
validate_dfxeta(xeredar::valid_data_xeta[["triac_japan_unspiked"]],
 row = FALSE)
}
