% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validators.R
\name{validate_mixedanova}
\alias{validate_mixedanova}
\title{Validate if model is as expected}
\usage{
validate_mixedanova(mixedaov)
}
\arguments{
\item{mixedaov}{model to validate}
}
\value{
In case the mixed ANOVA model is valid,
  the function invisibly returns it to the parent function
}
\description{
Validate if model is as expected
}
\examples{
mixedaov <- lme4::lmer(
  Fluor ~ Conc +
    (1 | Replicate) + (1 | Replicate:Conc) + (1 | Replicate:Conc:Row),
  data = xeredar::testDataSpiked, REML = TRUE
)
validate_mixedanova(mixedaov = mixedaov)
}
