% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XETAfunctions.R
\name{validity}
\alias{validity}
\title{validity: This function regards the validity criterium in the
guideline according to which the CV in each run and over all
runs is not suposed to be higher than 30\%.}
\usage{
validity(
  dataframe,
  dat_spiked_unspiked,
  row = TRUE,
  dagostino = FALSE,
  trimming = TRUE,
  outlier = TRUE,
  boxcox = TRUE,
  artc = FALSE,
  alpha = 0.01
)
}
\arguments{
\item{dataframe}{the raw dataframe that contains both spiked,
unspiked data, including the T4 control}

\item{dat_spiked_unspiked}{either the dataframe with the
spiked or unspiked samples}

\item{row}{Row Default is TRUE, which means that by default
the full model as specified in the Guideline is used. If it
is FALSE, the reduced model is selected whereby the
variability in the exposure wells is disregarded.}

\item{dagostino}{boolean. Default is FALSE; when FALSE,
the shapiro wilks test from the stats package is carried
out, when TRUE the dAgostina Pearson normality test from
the fBasics test is carried out.}

\item{trimming}{boolean. When TRUE, trimming is conducted by
the dataPrep function when required. When FALSE trimming
is never conducted. Default is TRUE.}

\item{outlier}{boolean, When TRUE, outlier removal is
conducted by the dataPrep function when required. When
FALSE, outlier removal is never conducted. Default is TRUE.}

\item{boxcox}{boolean, When TRUE boxcox transformation is
performed by the dataPrep function when necessary and
when no zeros are present in the data. When FALSE it will
never be conducted. Default is TRUE.}

\item{artc}{boolean, When TRUE ART-C transformation is
performed when necessary. When FALSE it will never be
conducted. Default is FALSE.
This feature is experimental!}

\item{alpha}{Numeric. Default is 0.01. This value is the
alpha value for the normality and variance homogeneity test.}
}
\value{
A list (collection) containing the trimmed or
 untrimmed data frame, two summary data frames containing
 information summarized based on the concentration and
 the replicate, the result from the shapiro and levene
 test, the boxplots per replicate, the result of the
 monotonicity test as well as the result from the
 Dunnett's and Williams test. Different from the
 output of the Ana function, the returned list also
 contains the mixed Anova model (mixedaov) whereby Fluor
 is transformed as recommended and the trimmed or
 untrimmed data was utilized. Furthermore, the output
 also contains a reasoning for the recommended
 transformation and trimming.
}
\description{
With this function the Coefficient of Variation
of the test medium control/ solvent control per run and
overall is determined and when it is over 30\% trimming is
conducted and it is decided whether data transformation is
necessary using the Ana() function. When the Coefficient of
Variation is smaller than 30\%, then this function calls the
data_prep() function.
}
\examples{
#data from the validity studies
Result1_1 <- validity(xeredar::valid_data_xeta[["triac_japan_combined"]],
 xeredar::valid_data_xeta[["triac_japan_spiked"]], row = FALSE)
Result1_2 <- validity(xeredar::valid_data_xeta[["triac_japan_combined"]],
 xeredar::valid_data_xeta[["triac_japan_unspiked"]], row = FALSE)

Result2_1 <- validity(xeredar::valid_data_xeta[["triac_france_combined"]],
 xeredar::valid_data_xeta[["triac_france_spiked"]], row = FALSE)
Result2_2 <- validity(xeredar::valid_data_xeta[["triac_france_combined"]],
 xeredar::valid_data_xeta[["triac_france_unspiked"]], row = FALSE)

Result3_1 <- validity(xeredar::valid_data_xeta[["triac_usa_combined"]],
 xeredar::valid_data_xeta[["triac_usa_spiked"]], row = FALSE)
Result3_2 <- validity(xeredar::valid_data_xeta[["triac_usa_combined"]],
 xeredar::valid_data_xeta[["triac_usa_unspiked"]], row = FALSE)
}
