% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCharrms_original.R
\name{williamsTest}
\alias{williamsTest}
\title{williamsTest: Performs Williams Test}
\usage{
williamsTest(df, resp, trt, direction = "decreasing", SeIn = NULL)
}
\arguments{
\item{df}{The data frame; Each row is an observation.}

\item{resp}{The name (as a string) of the response variable.}

\item{trt}{The name (as a string) of the treatment variable.}

\item{direction}{The direction of the test statistic which
can either be "increasing", "decreasing").}

\item{SeIn}{The Standard error, default is program selected
however WilliamsTest can take in a different value in the
case of repeated measures}
}
\value{
A dataframe with the following columns
 trt The treatment level the test statistic corresponds to.
 Y.Tilde The amalgamated averages for the treatment level
 Se_Diff The standard error.
 DF The degrees of freedom.
 Will The value of the Williams test statistic
 TCrit The critical value of the Williams test statistic,
 corresponding to a p-value of 0.05.
}
\description{
This function is originally written by
Joe Swintek and encodes the Williams test used in
ecotoxicology. It is recommended in most ecotoxicological
guidelines and the OECD Guidance Document No. 54 on
statistics in ecotoxicology. A detailed description
is provided in the book of John Green (Green, 2018).
}
\examples{
data(lengthWeightData)
SubData <- lengthWeightData[lengthWeightData$Age == "16 week", ]
SubData <- SubData[SubData$Generation == "F1", ]
SubData <- SubData[SubData$SEX == "M", ]
williamsTest(
  df = SubData, trt = "Treatment",
  resp = "WEIGHT", direction = "decreasing", SeIn = NULL
)
}
\references{
\url{https://www.rdocumentation.org/packages/StatCharrms/versions/0.90.96}

Green J.W., Springer T.A., Holbech H. (2018)
Statistical analysis of ecotoxicity studies.
First edition, John Wiley & Sons, Inc.

OECD (2006) Current approaches in the statistical
analysis of ecotoxicity data: A guidance to application.
Environmental Health and Safety Publications. Series on
Testing and Assessment. No. 54. Paris.

Williams D.A. (1971). A test for differences
between treatment means when several dose levels are
compared with a zero dose control. Biometrics 27(1):103-117.
}
