% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{williamsTestLookUpTable}
\alias{williamsTestLookUpTable}
\title{williamsTestLookUpTable from StatCharrms}
\format{
An object of class \code{data.frame} with 119 rows and 19 columns.
}
\usage{
williamsTestLookUpTable
}
\description{
This is the look up table for the critical values provided in
  Green (2018) derived from the StatCharrms package.
  This table is automatically called during the the execution of williamsTest
  function. Q and B: If c= number of reps on controls,
  r = number if reps in each treatment, w=c/r,
  then tcrit= Q- 10^-2 *B(1-1/w).
}
\keyword{datasets}
