% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RADAR_valid_data_table_spiked_unspiked}
\alias{RADAR_valid_data_table_spiked_unspiked}
\title{RADAR_valid_data_table_spiked_unspiked list of data used for testing the
functions.
It contains spiked and unspiked data from all substances and labs from the
RADAR validation studies.
For all datasets the column Conc is a factor with specified
levels, as the Williams test will otherwise not order the concentrations in
the correct way.}
\format{
An object of class \code{list} of length 44.
}
\usage{
RADAR_valid_data_table_spiked_unspiked
}
\description{
RADAR_valid_data_table_spiked_unspiked list of data used for testing the
functions.
It contains spiked and unspiked data from all substances and labs from the
RADAR validation studies.
For all datasets the column Conc is a factor with specified
levels, as the Williams test will otherwise not order the concentrations in
the correct way.
}
\keyword{datasets}
